chore: remove deprecated or unused files
Because random sampling in the end only was performed with RAVEN, these
scripts are removed from the repo to avoid confusion.
